Refactoring Eat might be ready to go
